+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this package will be documented in this file.
 
 - Changed language files
+- Added an icon button menu in **AppBar**
 
 # 0.5.0
 - Added mdx files for **overview** and **installation**